fix win compilation after r24267
TBR=marja@chromium.org

BUG=

Review URL: https://codereview.chromium.org/610373003

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24272 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/x64/instruction-selector-x64.cc
@@ -257,7 +257,7 @@ void InstructionSelector::VisitLoad(Node* node) {
   InstructionCode code = opcode | AddressingModeField::encode(matcher.mode_);
   InstructionOperand* outputs[] = {g.DefineAsRegister(node)};
   InstructionOperand* inputs[AddressingModeMatcher::kMaxInputCount];
-  int input_count = matcher.SetInputs(inputs);
+  size_t input_count = matcher.SetInputs(inputs);
   Emit(code, 1, outputs, input_count, inputs);
 }
 
@@ -319,7 +319,7 @@ void InstructionSelector::VisitStore(Node* node) {
   AddressingModeMatcher matcher(&g, base, index);
   InstructionCode code = opcode | AddressingModeField::encode(matcher.mode_);
   InstructionOperand* inputs[AddressingModeMatcher::kMaxInputCount + 1];
-  int input_count = matcher.SetInputs(inputs);
+  size_t input_count = matcher.SetInputs(inputs);
   inputs[input_count++] = val;
   Emit(code, 0, static_cast<InstructionOperand**>(NULL), input_count, inputs);
 }
